fix: Share single MSAL instance across auth modules
Previously AuthProvider.tsx and tokenService.ts each created their own
PublicClientApplication instance. This caused the login bug where users
were immediately logged out after authentication - the tokenService's
separate instance had no knowledge of the authenticated account.

Solution: Extract shared MSAL instance to msalInstance.ts module that
both files import, ensuring consistent authentication state.
